fix: refactor Terragrunt and Terraform (#7)
Update the folder structure to have the environment's Terragrunt
configurations reference a single `./aws` Terraform root module.

As part of this change, the Cognito and Google OAuth secrets will now
be created in both environments, but they will only have values in the
environments that are using them as identity providers.